requirements.txt
antlr4-python3-runtime==4.13.2
antlr4-tools==0.2.2
cssutils==2.11.1
more-itertools==10.8.0
